requirements.txt
anyio==3.6.1
boto3==1.25.4
botocore==1.28.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
et-xmlfile==1.1.0
fastapi==0.80.0
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.1
numpy==1.24.3
openai==0.25.0
openpyxl==3.1.2
pandas==2.0.2
pandas-stubs==2.0.2.230605
psycopg2==2.9.5
pydantic==1.9.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
sentry-sdk==1.14.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tqdm==4.65.0
types-pytz==2023.3.0.0
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
